[WJ] Rename go-server to go-parser
