+++ go-parser/main.go
@@ -3,8 +3,9 @@ package main
 import (
 	"fmt"
 	"net/http"
-	"github.com/xlzior/goats/backend/go-server/handler"
+
 	"github.com/gorilla/handlers"
+	"github.com/xlzior/goats/go-parser/handler"
 )
 
 func main() {
@@ -12,8 +13,8 @@ func main() {
 	// Define the endpoint handler functions
 	http.HandleFunc("/parse", handler.ParseHandler)
 
-  // CORS middleware configuration
-  corsHandler := handlers.CORS(
+	// CORS middleware configuration
+	corsHandler := handlers.CORS(
 		handlers.AllowedOrigins([]string{"*"}),
 		handlers.AllowedMethods([]string{"GET", "POST", "PUT", "DELETE", "OPTIONS"}),
 		handlers.AllowedHeaders([]string{"Content-Type", "Authorization"}),
